+++ lib/morphnext.dart
@@ -2,6 +2,7 @@
 library;
 
 export 'package:flutter/physics.dart' show SpringDescription;
+export 'src/cache/morph_cache.dart' show MorphCache;
 export 'src/morph_spring.dart' show MorphSprings;
 export 'src/widgets/animated_morph_icon.dart' show AnimatedMorphIcon;
 export 'src/widgets/morph_icon.dart' show MorphIcon;